v8.1.3: Render Alliance Member Manager in canonical Tools UI
