+++ .github/fixtures/issue553-alliance-member-manager-menu.json
@@ -1,42 +1,16 @@
 {
-  "schemaVersion": 2,
-  "description": "Issue #553 production menu labels independent of rendered card CSS classes.",
-  "cases": [
+  "schemaVersion": 3,
+  "description": "Canonical Tools control and persisted state cases for Issue #553.",
+  "states": [
     {
-      "name": "live rendered card with unrelated class",
-      "attributeIndex": null,
-      "labels": [
-        "Clean",
-        "Alliance Map Blocker",
-        "Personal Missions"
-      ],
-      "expectedIndex": 1
+      "enabled": false,
+      "pill": "OFF",
+      "pressed": "false"
     },
     {
-      "name": "feature attribute remains authoritative",
-      "attributeIndex": 0,
-      "labels": [
-        "Legacy attributed blocker",
-        "Alliance Map Blocker"
-      ],
-      "expectedIndex": 0
-    },
-    {
-      "name": "rendered whitespace is normalized",
-      "attributeIndex": null,
-      "labels": [
-        "  Alliance   Map\nBlocker  "
-      ],
-      "expectedIndex": 0
-    },
-    {
-      "name": "similar labels do not false-match",
-      "attributeIndex": null,
-      "labels": [
-        "Alliance Map Block",
-        "Alliance Member Manager"
-      ],
-      "expectedIndex": null
+      "enabled": true,
+      "pill": "ON",
+      "pressed": "true"
     }
   ]
 }
