+++ .github/fixtures/main-style-source-headroom.json
@@ -102,9 +102,9 @@
   "v8Candidate": {
     "issue": 553,
     "version": "8.1.3",
-    "sourceBytes": 1644850,
-    "sourceLines": 25076,
-    "sourceSha256": "af78bf9720981dacf7d89454dd323a2627250d03e8092e5324ef1b47c6586307",
+    "sourceBytes": 1639133,
+    "sourceLines": 24945,
+    "sourceSha256": "5524d0f2181d9d7a0d0affa401546b729862c06aa81c2f596b59f11d5d706c95",
     "templateBytes": 683558,
     "templateLines": 7858,
     "templateSha256": "8be4b32d5d4a001e15a5606ac28238fef22909e3276d791a3350851611602226",
@@ -114,11 +114,11 @@
     "minTemplateBytes": 500000,
     "baseline": "8.1.2",
     "approvedGrowth": {
-      "sourceBytes": 2927,
-      "sourceLines": 64,
+      "sourceBytes": -2790,
+      "sourceLines": -67,
       "templateBytes": 0,
       "templateLines": 0
     },
-    "scope": "Issue #553 native Alliance Member Manager Tools card cloning, class-independent lookup and panel-scoped re-render recovery"
+    "scope": "Issue #553 canonical createPanel Alliance Member Manager control, panel-owned action routing and normal updateUI state reconciliation"
   }
 }
