+++ .github/performance-budget.json
@@ -1,12 +1,12 @@
 {
   "schemaVersion": 1,
-  "revision": "2026-07-27-issue-553-native-menu-render",
-  "rationale": "Issue #553 renders the Alliance Member Manager from the canonical live Tools card and adds one panel-scoped child-list observer with coalesced bounded frame reconciliation; the member-page network and runtime budgets remain unchanged.",
+  "revision": "2026-07-27-issue-553-canonical-menu-render",
+  "rationale": "Issue #553 renders Alliance Member Manager directly in createPanel beside Alliance Map Blocker, routes clicks through the existing panel action handler and adds no observer, timer, animation-frame loop or network site.",
   "transitionApproval": {
     "issue": 553,
     "version": "8.1.3",
     "approvedNetworkRequestDelta": 1,
-    "scope": "Native Alliance Member Manager Tools rendering with one panel-scoped observer, no timer loop and unchanged user-triggered sequential fetch ownership."
+    "scope": "Canonical Alliance Member Manager Tools markup and panel-owned action routing; the existing user-triggered sequential member-page fetch site is unchanged."
   },
   "requiredSourceFragments": [
     "@run-at       document-start",
