+++ .github/scripts/test_alliance_member_manager_contract.py
@@ -28,7 +28,6 @@ def main() -> int:
     assert "version: '8.1.3'" in source
     for marker in [
         "mcms_alliance_member_manager_enabled_v1",
-        "Alliance Operations",
         "Alliance Member Manager",
         "Role",
         "Activity",
@@ -44,16 +43,6 @@ def main() -> int:
         "Member name",
         "Alliance role",
         "Showing ${visible} of ${context.members.size} members",
-        "allianceBuildingsMapBlocker",
-        "panel.querySelectorAll('.mcms-label')",
-        "allianceMemberManagerRenderedLabel(label) === 'Alliance Map Blocker'",
-        "button = blocker.cloneNode(true)",
-        "attribute.startsWith('data-')",
-        "Alliance Operations",
-        "allianceMemberManagerMenuObserver.observe(panel, { childList: true, subtree: true })",
-        "requestAnimationFrame",
-        "data-mcms-alliance-member-manager-toggle",
-        "data-mcms-alliance-operations",
         r"alliance\/members|verband\/mitglieder",
         "img.online_icon",
         "user_(?<state>blue|gray|green|red|yellow)",
@@ -67,7 +56,7 @@ def main() -> int:
         "context.originalRows.forEach(row =>",
         "typeof document.addEventListener === 'function'",
         "document.readyState !== 'loading'",
-        "document.addEventListener('DOMContentLoaded'",
+        "document.addEventListener('DOMContentLoaded', reconcileAllianceMemberManager",
         "{ once: true }",
         "allianceMemberManagerOtherOwnerPresent()",
         "#allianceMemberList-controls",
@@ -76,6 +65,12 @@ def main() -> int:
 
     for forbidden in [
         "setInterval(",
+        "new MutationObserver(",
+        "requestAnimationFrame(",
+        "allianceMemberManagerMapBlockerButton",
+        "ensureAllianceMemberManagerMenuControl",
+        "queueAllianceMemberManagerMenuControl",
+        "allianceMemberManagerMenuObserver",
         "GM_xmlhttpRequest",
         "last seen",
         "lastSeen",
@@ -83,25 +78,36 @@ def main() -> int:
     ]:
         assert forbidden not in block, forbidden
 
+    for marker in [
+        "function makeAllianceMemberManagerToggleButton()",
+        'data-action="toggle-alliance-member-manager"',
+        'data-mcms-alliance-member-manager-toggle="true"',
+        '<div class="mcms-section-label" data-mcms-alliance-operations="label">Alliance Operations</div>',
+        '<div class="mcms-grid-2" data-mcms-alliance-operations="controls">',
+        "${makeAllianceMemberManagerToggleButton()}",
+        "if (action === 'toggle-alliance-member-manager')",
+        "setAllianceMemberManagerEnabled(!allianceMemberManagerEnabled())",
+        "updateAllianceMemberManagerMenuControl();",
+    ]:
+        assert marker in source, marker
+
     assert block.index("for (let page = 1;") < block.index("await fetch(")
     assert block.count("await fetch(") == 1
     assert block.count("new AbortController()") == 1
-    assert block.count("new MutationObserver(") == 1
-    assert ".mcms-toggle-btn" not in block
-    assert "setInterval(" not in block
+    assert block.count("new MutationObserver(") == 0
     assert block.count("data-mcms-alliance-member-manager-toggle") == 1
-    assert block.count("Alliance Operations") == 1
+    assert source.count("${makeAllianceMemberManagerToggleButton()}") == 1
+    assert source.count('data-action="toggle-alliance-member-manager"') == 1
 
     changelog = CHANGELOG.read_text(encoding="utf-8")
     assert "## [8.1.3] - 2026-07-27" in changelog
-    assert "### Native Alliance Member Manager Tools rendering" in changelog
+    assert "### Canonical Alliance Member Manager Tools rendering" in changelog
     assert "## [8.1.2] - 2026-07-27" in changelog
     assert "## [8.1.1] - 2026-07-27" in changelog
     assert "## [8.1.0] - 2026-07-27" in changelog
 
     site_data = json.loads(SITE_DATA.read_text(encoding="utf-8"))
-    categories = site_data["featureCategories"]
-    alliance = [item for item in categories if item.get("name") == "Alliance operations"]
+    alliance = [item for item in site_data["featureCategories"] if item.get("name") == "Alliance operations"]
     assert len(alliance) == 1
     features = alliance[0]["features"]
     assert len(features) == 1
@@ -114,17 +120,15 @@ def main() -> int:
     assert "test_issue554_alliance_member_manager_rollback.py" not in preflight
 
     performance = json.loads(PERFORMANCE.read_text(encoding="utf-8"))
-    assert performance["revision"] == "2026-07-27-issue-553-native-menu-render"
+    assert performance["revision"] == "2026-07-27-issue-553-canonical-menu-render"
     assert performance["transitionApproval"]["issue"] == 553
     assert performance["transitionApproval"]["version"] == "8.1.3"
     assert performance["transitionApproval"]["approvedNetworkRequestDelta"] == 1
     assert performance["absoluteLimits"]["network_request_calls"] == 5
-    assert performance["relativeLimits"]["network_request_calls"]["warnDelta"] == 1
-    assert performance["relativeLimits"]["network_request_calls"]["failDelta"] == 1
 
     print(
-        "Alliance Member Manager contract passed: English controls, explicit sequential loading, "
-        "current-state filtering, cloned native Tools control, panel-scoped render reconciliation and deterministic teardown."
+        "Alliance Member Manager contract passed: canonical Tools markup, panel-owned action routing, "
+        "persisted state reconciliation, responsive member controls and zero added observers."
     )
     return 0
