+++ .github/scripts/test_issue553_alliance_member_manager_menu_runtime.js
@@ -36,53 +36,64 @@ function extractFunction(name) {
   throw new Error(`Unable to extract ${name}`);
 }
 
-const functionText = [
-  extractFunction("allianceMemberManagerRenderedLabel"),
-  extractFunction("allianceMemberManagerMapBlockerButton"),
-].join("\n");
-const sandbox = {};
-vm.runInNewContext(
-  `${functionText}\nthis.resolveBlocker = allianceMemberManagerMapBlockerButton;`,
-  sandbox
-);
+const helperText = extractFunction("makeAllianceMemberManagerToggleButton");
+const helperSandbox = {};
+vm.runInNewContext(`${helperText}\nthis.renderManagerButton = makeAllianceMemberManagerToggleButton;`, helperSandbox);
+const html = helperSandbox.renderManagerButton();
+for (const marker of [
+  'class="mcms-toggle-btn"',
+  'data-action="toggle-alliance-member-manager"',
+  'data-mcms-alliance-member-manager-toggle="true"',
+  'aria-pressed="false"',
+  '<span class="mcms-iconbox">AM</span>',
+  '<span class="mcms-label">Alliance Member Manager</span>',
+  '<span class="mcms-pill">OFF</span>',
+]) assert.ok(html.includes(marker), marker);
 
-for (const item of fixture.cases) {
-  const cards = item.labels.map((textContent, index) => ({ index, kind: "canonical-live-card" }));
-  const labels = item.labels.map((textContent, index) => ({
-    textContent,
-    closest(selector) {
-      assert.equal(selector, 'button, a, [role="button"], [tabindex]');
-      return cards[index];
-    },
-    parentElement: null,
-  }));
-  const panel = {
-    querySelector(selector) {
-      assert.match(selector, /allianceBuildingsMapBlocker/);
-      return item.attributeIndex === null ? null : cards[item.attributeIndex];
-    },
-    querySelectorAll(selector) {
-      assert.equal(selector, ".mcms-label");
-      return labels;
-    },
-  };
-  const result = sandbox.resolveBlocker(panel);
-  if (item.expectedIndex === null) assert.equal(result, null, item.name);
-  else assert.equal(result, cards[item.expectedIndex], item.name);
+for (const marker of [
+  '<div class="mcms-section-label" data-mcms-alliance-operations="label">Alliance Operations</div>',
+  '<div class="mcms-grid-2" data-mcms-alliance-operations="controls">',
+  "${makeToggleButton('allianceBuildingsMapBlocker'",
+  "${makeAllianceMemberManagerToggleButton()}",
+  "if (action === 'toggle-alliance-member-manager')",
+  "setAllianceMemberManagerEnabled(!allianceMemberManagerEnabled())",
+]) assert.ok(source.includes(marker), marker);
+
+const updateText = extractFunction("updateAllianceMemberManagerMenuControl");
+const buttonState = { on: false, pressed: "false", pill: "OFF" };
+const pill = { get textContent() { return buttonState.pill; }, set textContent(value) { buttonState.pill = value; } };
+const button = {
+  classList: { toggle(name, value) { assert.equal(name, "mcms-on"); buttonState.on = value; } },
+  setAttribute(name, value) { if (name === "aria-pressed") buttonState.pressed = value; },
+  querySelector(selector) { assert.equal(selector, ".mcms-pill"); return pill; },
+};
+const panel = { querySelector(selector) { assert.equal(selector, "[data-mcms-alliance-member-manager-toggle]"); return button; } };
+const sandbox = {
+  SCRIPT: { panelId: "toolkit-panel" },
+  ALLIANCE_MEMBER_MANAGER: { menuAttribute: "data-mcms-alliance-member-manager-toggle" },
+  document: { querySelector(selector) { assert.equal(selector, "#toolkit-panel"); return panel; } },
+  enabled: false,
+};
+sandbox.allianceMemberManagerEnabled = () => sandbox.enabled;
+vm.runInNewContext(`${updateText}\nthis.updateManagerButton = updateAllianceMemberManagerMenuControl;`, sandbox);
+for (const item of fixture.states) {
+  sandbox.enabled = item.enabled;
+  sandbox.updateManagerButton();
+  assert.equal(buttonState.on, item.enabled);
+  assert.equal(buttonState.pressed, item.pressed);
+  assert.equal(buttonState.pill, item.pill);
 }
 
 const managerStart = source.indexOf("    // <mcms-alliance-member-manager>");
 const managerEnd = source.indexOf("    // </mcms-alliance-member-manager>", managerStart);
 const manager = source.slice(managerStart, managerEnd);
-for (const marker of [
-  "button = blocker.cloneNode(true)",
-  "attribute.startsWith('data-')",
-  "sectionLabel.textContent = 'Alliance Operations'",
-  "allianceMemberManagerMenuObserver.observe(panel, { childList: true, subtree: true })",
-  "requestAnimationFrame",
-  "group.append(button)",
-]) assert.ok(manager.includes(marker), marker);
-assert.equal((manager.match(/new MutationObserver\(/g) || []).length, 1);
-assert.ok(!manager.includes("panel.querySelectorAll('.mcms-toggle-btn')"));
+for (const forbidden of [
+  "new MutationObserver(",
+  "requestAnimationFrame(",
+  "allianceMemberManagerMapBlockerButton",
+  "ensureAllianceMemberManagerMenuControl",
+  "queueAllianceMemberManagerMenuControl",
+  "allianceMemberManagerMenuObserver",
+]) assert.ok(!manager.includes(forbidden), forbidden);
 
-console.log(`Issue #553 native menu runtime passed: ${fixture.cases.length} class-independent cases plus clone and re-render ownership.`);
+console.log(`Issue #553 canonical Tools runtime passed: ${fixture.states.length} persisted states and zero post-render injection.`);
