+++ help/manifest.json
@@ -15,6 +15,6 @@
   "financialIntelligenceRules": "financial-intelligence/v1/classification-rules.json",
   "financialAuditPolicy": "financial-intelligence/v2/audit-policy.json",
   "financialArchitectureManifest": "financial-intelligence/v2/manifest.json",
-  "runtimeGuidePatch": "Toolkit v8.1.3 renders the Alliance Member Manager from the real live Tools card, keeps it beside Alliance Map Blocker under Alliance Operations and restores it after normal panel re-renders.",
+  "runtimeGuidePatch": "Toolkit v8.1.3 renders Alliance Member Manager directly in the canonical Tools panel beside Alliance Map Blocker under Alliance Operations, with panel-owned clicks and normal persisted ON/OFF state reconciliation.",
   "godfatherThemePackage": "themes/godfather/manifest.json"
 }
